refactor step "I have created {int} categories"
