+++ e2e/src/domains/category/application/use-cases/index.ts
@@ -10,6 +10,7 @@ export { SetupMultipleCategoriesUseCase } from "./setup-multiple-categories.use-
 export { SetupBothCategoryTypesUseCase } from "./setup-both-category-types.use-case";
 export { SetupMultipleTestCategoriesUseCase } from "./setup-multiple-test-categories.use-case";
 export { SetupMultipleIncomeExpenseCategoriesUseCase } from "./setup-multiple-income-expense-categories.use-case";
+export { SetupBulkCategoriesUseCase } from "./setup-bulk-categories.use-case";
 export { NavigateToApplicationUseCase } from "./navigate-to-application.use-case";
 
 // Search use cases
